fix empty lines and move some function to RsExprUtils
