+++ src/main/kotlin/org/rust/lang/utils/RsDiagnostic.kt
@@ -767,7 +767,6 @@ sealed class RsDiagnostic(
         )
     }
 
-
     class ExperimentalFeature(
         element: PsiElement,
         private val presentableFeatureName: String,
